Fix compiler warning on pushbuild on 32 bit Windows

+++ storage/archive/ha_archive.cc
@@ -658,7 +658,7 @@ int ha_archive::create(const char *name, TABLE *table_arg,
         frm_ptr= (uchar *)my_malloc(sizeof(uchar) * (size_t)file_stat.st_size, MYF(0));
         if (frm_ptr)
         {
-          my_read(frm_file, frm_ptr, file_stat.st_size, MYF(0));
+          my_read(frm_file, frm_ptr, (size_t)file_stat.st_size, MYF(0));
           azwrite_frm(&create_stream, (char *)frm_ptr, file_stat.st_size);
           my_free((uchar*)frm_ptr, MYF(0));
         }
